requirements.txt
Django==5.0.1
django_channels==0.7.0
redis
daphne
